# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type MembershipRole = typeof MembershipRole[keyof typeof MembershipRole];


export const MembershipRole = {
  agency_admin: 'agency_admin',
  sub_account_holder: 'sub_account_holder',
  member: 'member',
} as const;

export type MembershipStatus = typeof MembershipStatus[keyof typeof MembershipStatus];


export const MembershipStatus = {
  active: 'active',
  revoked: 'revoked',
} as const;

export interface Membership {
  id: string;
  userId: string;
  agencyId: string;
  /** @nullable */
  subAccountId?: string | null;
  role: MembershipRole;
  status: MembershipStatus;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  accessExpiresAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Me {
  userId: string;
  memberships: Membership[];
}

export type EntitlementReason = typeof EntitlementReason[keyof typeof EntitlementReason];


export const EntitlementReason = {
  active: 'active',
  none: 'none',
  revoked: 'revoked',
  expired: 'expired',
} as const;

export interface Entitlement {
  active: boolean;
  reason: EntitlementReason;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  agencyId?: string | null;
  /** @nullable */
  subAccountId?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface SubAccount {
  id: string;
  agencyId: string;
  name: string;
  memberCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface SubAccountInput {
  /** @minLength 1 */
  name: string;
}

export type InviteRole = typeof InviteRole[keyof typeof InviteRole];


export const InviteRole = {
  agency_admin: 'agency_admin',
  sub_account_holder: 'sub_account_holder',
  member: 'member',
} as const;

export type InviteStatus = typeof InviteStatus[keyof typeof InviteStatus];


export const InviteStatus = {
  pending: 'pending',
  redeemed: 'redeemed',
  revoked: 'revoked',
} as const;

export interface Invite {
  id: string;
  code: string;
  agencyId: string;
  /** @nullable */
  subAccountId?: string | null;
  role: InviteRole;
  /** @nullable */
  accessDurationDays?: number | null;
  /** @nullable */
  inviteExpiresAt?: string | null;
  status: InviteStatus;
  createdByUserId: string;
  /** @nullable */
  redeemedByUserId?: string | null;
  /** @nullable */
  redeemedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type InviteInputRole = typeof InviteInputRole[keyof typeof InviteInputRole];


export const InviteInputRole = {
  sub_account_holder: 'sub_account_holder',
  member: 'member',
} as const;

export interface InviteInput {
  subAccountId: string;
  role: InviteInputRole;
  /** @minimum 1 */
  accessDurationDays?: number;
  /** @minimum 1 */
  inviteExpiresInDays?: number;
}

export interface InviteRedemption {
  /** @minLength 1 */
  code: string;
}

export interface AccessRenewal {
  /** @minimum 1 */
  accessDurationDays?: number;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Not permitted
 */
export type ForbiddenResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Conflicting state
 */
export type ConflictResponse = Error;

export type ListInvitesParams = {
subAccountId?: string;
};

export type ListMembersParams = {
subAccountId?: string;
};