# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: access
    description: Identity, entitlement, agencies and sub-accounts
  - name: invites
    description: Invite creation and redemption
  - name: members
    description: Member access management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /me:
    get:
      operationId: getMe
      tags: [access]
      summary: Current user's identity and memberships
      responses:
        "200":
          description: The authenticated user's memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /me/entitlement:
    get:
      operationId: getEntitlement
      tags: [access]
      summary: Server-authoritative access entitlement for the current user
      responses:
        "200":
          description: The user's current entitlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlement"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /sub-accounts:
    get:
      operationId: listSubAccounts
      tags: [access]
      summary: List sub-accounts in scope
      responses:
        "200":
          description: Sub-accounts the caller can manage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createSubAccount
      tags: [access]
      summary: Create a sub-account (agency admins only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /invites:
    get:
      operationId: listInvites
      tags: [invites]
      summary: List invites in scope
      parameters:
        - name: subAccountId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Invites the caller can manage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createInvite
      tags: [invites]
      summary: Create an invite for a sub-account holder or member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /invites/redeem:
    post:
      operationId: redeemInvite
      tags: [invites]
      summary: Redeem an invite code for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteRedemption"
      responses:
        "201":
          description: Membership created from the invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"

  /invites/{id}/revoke:
    post:
      operationId: revokeInvite
      tags: [invites]
      summary: Revoke a pending invite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The revoked invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members in scope
      parameters:
        - name: subAccountId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Members the caller can manage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Membership"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /members/{id}/renew:
    post:
      operationId: renewMember
      tags: [members]
      summary: Renew a member's access window
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRenewal"
      responses:
        "200":
          description: The updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /members/{id}/revoke:
    post:
      operationId: revokeMember
      tags: [members]
      summary: Revoke a member's access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflicting state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Membership:
      type: object
      required:
        - id
        - userId
        - agencyId
        - role
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        agencyId:
          type: string
          format: uuid
        subAccountId:
          type: ["string", "null"]
          format: uuid
        role:
          type: string
          enum: [agency_admin, sub_account_holder, member]
        status:
          type: string
          enum: [active, revoked]
        email:
          type: ["string", "null"]
        accessExpiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Me:
      type: object
      required:
        - userId
        - memberships
      properties:
        userId:
          type: string
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/Membership"

    Entitlement:
      type: object
      required:
        - active
        - reason
      properties:
        active:
          type: boolean
        reason:
          type: string
          enum: [active, none, revoked, expired]
        role:
          type: ["string", "null"]
        agencyId:
          type: ["string", "null"]
        subAccountId:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time

    SubAccount:
      type: object
      required:
        - id
        - agencyId
        - name
        - memberCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        agencyId:
          type: string
          format: uuid
        name:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SubAccountInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1

    Invite:
      type: object
      required:
        - id
        - code
        - agencyId
        - role
        - status
        - createdByUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        agencyId:
          type: string
          format: uuid
        subAccountId:
          type: ["string", "null"]
          format: uuid
        role:
          type: string
          enum: [agency_admin, sub_account_holder, member]
        accessDurationDays:
          type: ["integer", "null"]
        inviteExpiresAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, redeemed, revoked]
        createdByUserId:
          type: string
        redeemedByUserId:
          type: ["string", "null"]
        redeemedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    InviteInput:
      type: object
      required:
        - subAccountId
        - role
      properties:
        subAccountId:
          type: string
          format: uuid
        role:
          type: string
          enum: [sub_account_holder, member]
        accessDurationDays:
          type: integer
          minimum: 1
        inviteExpiresInDays:
          type: integer
          minimum: 1

    InviteRedemption:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          minLength: 1

    AccessRenewal:
      type: object
      properties:
        accessDurationDays:
          type: integer
          minimum: 1