# lib/api-zod/src/generated/types/entitlement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EntitlementReason } from './entitlementReason';

export interface Entitlement {
  active: boolean;
  reason: EntitlementReason;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  agencyId?: string | null;
  /** @nullable */
  subAccountId?: string | null;
  /** @nullable */
  expiresAt?: Date | null;
}