# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InviteRole } from './inviteRole';
import type { InviteStatus } from './inviteStatus';

export interface Invite {
  id: string;
  code: string;
  agencyId: string;
  /** @nullable */
  subAccountId?: string | null;
  role: InviteRole;
  /** @nullable */
  accessDurationDays?: number | null;
  /** @nullable */
  inviteExpiresAt?: Date | null;
  status: InviteStatus;
  createdByUserId: string;
  /** @nullable */
  redeemedByUserId?: string | null;
  /** @nullable */
  redeemedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}