# lib/api-zod/src/generated/types/inviteRedemption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InviteRedemption {
  /** @minLength 1 */
  code: string;
}